requirements.txt
annotated-types==0.7.0
anyio==4.4.0
beanie==1.26.0
click==8.1.7
dnspython==2.6.1
exceptiongroup==1.2.2
fastapi==0.114.0
greenlet==3.0.3
h11==0.14.0
httptools==0.6.1
idna==3.8
lazy-model==0.2.0
motor==3.5.1
pip==22.0.2
pydantic==2.9.1
pydantic_core==2.23.3
pydantic-settings==2.4.0
pymongo==4.8.0
PyMySQL==1.1.1
python-decouple==3.8
python-dotenv==1.0.1
PyYAML==6.0.2
setuptools==59.6.0
sniffio==1.3.1
SQLAlchemy==2.0.34
sqlmodel==0.0.22
starlette==0.38.5
toml==0.10.2
typing_extensions==4.12.2
uvicorn==0.30.6
uvloop==0.20.0
watchfiles==0.24.0
websockets==13.0.1
